feat(engine): Propagate splitter uncertainty in MC simulation
This commit fixes a bug where uncertainty defined for a Splitter process's material TC was not being applied during Monte Carlo simulations. The run_mc_simulation function has been modified to ensure that when a value is sampled for a ..._material TC, it is correctly propagated to the corresponding ..._WC, ..._DM, and ..._CC parameters used by the solver. Additionally, this commit adds documentation for a deeper, underlying bug related to inconsistent TC naming conventions between the data loader and the solver, which will need to be addressed in a future session.